Refactored modals and embeds.
