+++ events/poll/newProposalPoll.js
@@ -2,7 +2,7 @@
 const { EmbedBuilder, TextChannel } = require('discord.js');
 const { Types } = require('mongoose');
 
-const { initPollMessage } = require('../../helpers/poll/initPollMessage');
+const { generateInitialPollMessage } = require('../../views/embeds/polls');
 const PollChannel = require('../../db/schemas/PollChannel');
 const PollCount = require('../../db/schemas/ChannelPollCount');
 const Poll = require('../../db/schemas/Poll');
@@ -42,7 +42,6 @@ module.exports = {
          client,
          guildId,
          guild: {
-            channels: { cache },
             roles: {
                everyone: { id: everyoneId },
             },
@@ -52,7 +51,6 @@ module.exports = {
             user,
             user: { username, discriminator },
          },
-         // } = interaction;
       } = interaction;
 
       Logger.info('events/poll/newProposalPoll.js: Creating new proposal.', {
@@ -105,7 +103,7 @@ module.exports = {
          allowedRoles: channelConfig.allowedRoles,
       });
 
-      const messageObject = await initPollMessage({
+      const messageObject = await generateInitialPollMessage({
          propId,
          title,
          description,
@@ -133,7 +131,6 @@ module.exports = {
             .then(fetchedMembers => {
                return fetchedMembers.filter(member => {
                   return (
-                     //    member.presence?.status === 'online' && //disabled not worrying about the online presence
                      !member.user.bot &&
                      member?.roles.cache.hasAny(...channelConfig.allowedRoles)
                   );
@@ -184,9 +181,7 @@ module.exports = {
             _id: new Types.ObjectId(),
             guildId,
             creatorId: user.id,
-            // messageId: message.id,
             messageId: interaction.id,
-            // config: config._id,
             config: channelConfig._id,
             pollData,
             votes: undefined,
@@ -202,7 +197,7 @@ module.exports = {
             .then(async poll => {
                await pollNumber.increment();
                poll.pollNumber = pollNumber.pollsCreated;
-               return await poll.save();
+               return poll.save();
             });
 
          const updateVoterPromise = [...newPoll.allowedUsers.keys()].map(
@@ -256,22 +251,13 @@ module.exports = {
                }
 
                user.markModified('eligibleChannels');
-               return await user.save();
+               return user.save();
             },
          );
 
          await Promise.all(updateVoterPromise);
 
          const updatedEmbed = new EmbedBuilder(messageObject.embeds[0]);
-
-         // const timeEndMilli = new Date(
-         //    newPoll.timeCreated.getTime() + durationMs
-
-         //    // !testing switching the time for testing purposes
-         //    // savedPoll.timeCreated.getTime() + 30000
-         // );
-         // newPoll.timeEnd = timeEndMilli.toISOString();
-         // await newPoll.save();
 
          updatedEmbed.setFooter({
             text: `Poll #${newPoll.pollNumber} submitted by ${
